Remove opt, dict, more
