+++ Common/Common.vcxproj
@@ -393,8 +393,6 @@
     <ClInclude Include="..\ext\at3_standalone\attributes.h" />
     <ClInclude Include="..\ext\at3_standalone\avcodec.h" />
     <ClInclude Include="..\ext\at3_standalone\avfft.h" />
-    <ClInclude Include="..\ext\at3_standalone\avstring.h" />
-    <ClInclude Include="..\ext\at3_standalone\avutil.h" />
     <ClInclude Include="..\ext\at3_standalone\bswap.h" />
     <ClInclude Include="..\ext\at3_standalone\buffer.h" />
     <ClInclude Include="..\ext\at3_standalone\buffer_internal.h" />
@@ -403,7 +401,6 @@
     <ClInclude Include="..\ext\at3_standalone\common.h" />
     <ClInclude Include="..\ext\at3_standalone\compat.h" />
     <ClInclude Include="..\ext\at3_standalone\config.h" />
-    <ClInclude Include="..\ext\at3_standalone\dict.h" />
     <ClInclude Include="..\ext\at3_standalone\error.h" />
     <ClInclude Include="..\ext\at3_standalone\fft-internal.h" />
     <ClInclude Include="..\ext\at3_standalone\fft.h" />
@@ -420,7 +417,6 @@
     <ClInclude Include="..\ext\at3_standalone\mathematics.h" />
     <ClInclude Include="..\ext\at3_standalone\mathops.h" />
     <ClInclude Include="..\ext\at3_standalone\mem.h" />
-    <ClInclude Include="..\ext\at3_standalone\opt.h" />
     <ClInclude Include="..\ext\at3_standalone\rational.h" />
     <ClInclude Include="..\ext\at3_standalone\samplefmt.h" />
     <ClInclude Include="..\ext\at3_standalone\sinewin.h" />
@@ -632,11 +628,10 @@
     <ClCompile Include="..\ext\at3_standalone\atrac3plusdsp.c" />
     <ClCompile Include="..\ext\at3_standalone\avfft.c" />
     <ClCompile Include="..\ext\at3_standalone\avpacket.c" />
-    <ClCompile Include="..\ext\at3_standalone\avstring.c" />
     <ClCompile Include="..\ext\at3_standalone\av_buffer.c" />
     <ClCompile Include="..\ext\at3_standalone\bitstream.c" />
     <ClCompile Include="..\ext\at3_standalone\channel_layout.c" />
-    <ClCompile Include="..\ext\at3_standalone\dict.c" />
+    <ClCompile Include="..\ext\at3_standalone\compat.c" />
     <ClCompile Include="..\ext\at3_standalone\fft_template.c" />
     <ClCompile Include="..\ext\at3_standalone\float_dsp.c" />
     <ClCompile Include="..\ext\at3_standalone\frame.c" />
@@ -644,7 +639,6 @@
     <ClCompile Include="..\ext\at3_standalone\mathematics.c" />
     <ClCompile Include="..\ext\at3_standalone\mdct_template.c" />
     <ClCompile Include="..\ext\at3_standalone\mem.c" />
-    <ClCompile Include="..\ext\at3_standalone\opt.c" />
     <ClCompile Include="..\ext\at3_standalone\options.c" />
     <ClCompile Include="..\ext\at3_standalone\rational.c" />
     <ClCompile Include="..\ext\at3_standalone\samplefmt.c" />
@@ -916,7 +910,6 @@
     </ClCompile>
     <ClCompile Include="ArmEmitter.cpp" />
     <ClCompile Include="Buffer.cpp" />
-    <ClCompile Include="compat.c" />
     <ClCompile Include="Data\Collections\FastVec.h" />
     <ClCompile Include="Data\Color\RGBAUtil.cpp" />
     <ClCompile Include="Data\Convert\SmallDataConvert.cpp" />
