+++ ext/at3_standalone/compat.c
@@ -18,6 +18,18 @@ void ff_fft_init_arm(void) {}
 void ff_fft_init_mips(void) {}
 void ff_fft_init_ppc(void) {}
 
+
+size_t av_strlcpy(char *dst, const char *src, size_t size)
+{
+	size_t len = 0;
+	while (++len < size && *src)
+		*dst++ = *src++;
+	if (len <= size)
+		*dst = 0;
+	return len + strlen(src) - 1;
+}
+
+
 const uint8_t ff_reverse[256] = {
 0x00,0x80,0x40,0xC0,0x20,0xA0,0x60,0xE0,0x10,0x90,0x50,0xD0,0x30,0xB0,0x70,0xF0,
 0x08,0x88,0x48,0xC8,0x28,0xA8,0x68,0xE8,0x18,0x98,0x58,0xD8,0x38,0xB8,0x78,0xF8,
