+++ ext/at3_standalone/frame.c
@@ -163,103 +163,6 @@ FF_ENABLE_DEPRECATION_WARNINGS
     return 0;
 }
 
-int av_frame_ref(AVFrame *dst, const AVFrame *src)
-{
-    int i, ret = 0;
-
-    dst->format         = src->format;
-    dst->channels       = src->channels;
-    dst->channel_layout = src->channel_layout;
-    dst->nb_samples     = src->nb_samples;
-
-    ret = frame_copy_props(dst, src, 0);
-    if (ret < 0)
-        return ret;
-
-    /* duplicate the frame data if it's not refcounted */
-    if (!src->buf[0]) {
-        ret = av_frame_get_buffer(dst, 32);
-        if (ret < 0)
-            return ret;
-
-        ret = av_frame_copy(dst, src);
-        if (ret < 0)
-            av_frame_unref(dst);
-
-        return ret;
-    }
-
-    /* ref the buffers */
-    for (i = 0; i < FF_ARRAY_ELEMS(src->buf); i++) {
-        if (!src->buf[i])
-            continue;
-        dst->buf[i] = av_buffer_ref(src->buf[i]);
-        if (!dst->buf[i]) {
-            ret = AVERROR(ENOMEM);
-            goto fail;
-        }
-    }
-
-    if (src->extended_buf) {
-        dst->extended_buf = av_mallocz_array(sizeof(*dst->extended_buf),
-                                       src->nb_extended_buf);
-        if (!dst->extended_buf) {
-            ret = AVERROR(ENOMEM);
-            goto fail;
-        }
-        dst->nb_extended_buf = src->nb_extended_buf;
-
-        for (i = 0; i < src->nb_extended_buf; i++) {
-            dst->extended_buf[i] = av_buffer_ref(src->extended_buf[i]);
-            if (!dst->extended_buf[i]) {
-                ret = AVERROR(ENOMEM);
-                goto fail;
-            }
-        }
-    }
-
-    /* duplicate extended data */
-    if (src->extended_data != src->data) {
-        int ch = src->channels;
-
-        if (!ch) {
-            ret = AVERROR(EINVAL);
-            goto fail;
-        }
-        CHECK_CHANNELS_CONSISTENCY(src);
-
-        dst->extended_data = av_malloc_array(sizeof(*dst->extended_data), ch);
-        if (!dst->extended_data) {
-            ret = AVERROR(ENOMEM);
-            goto fail;
-        }
-        memcpy(dst->extended_data, src->extended_data, sizeof(*src->extended_data) * ch);
-    } else
-        dst->extended_data = dst->data;
-
-    memcpy(dst->data,     src->data,     sizeof(src->data));
-    memcpy(dst->linesize, src->linesize, sizeof(src->linesize));
-
-    return 0;
-
-fail:
-    av_frame_unref(dst);
-    return ret;
-}
-
-AVFrame *av_frame_clone(const AVFrame *src)
-{
-    AVFrame *ret = av_frame_alloc();
-
-    if (!ret)
-        return NULL;
-
-    if (av_frame_ref(ret, src) < 0)
-        av_frame_free(&ret);
-
-    return ret;
-}
-
 void av_frame_unref(AVFrame *frame)
 {
     int i;
@@ -274,99 +177,4 @@ void av_frame_unref(AVFrame *frame)
     av_freep(&frame->extended_buf);
 
     get_frame_defaults(frame);
-}
-
-void av_frame_move_ref(AVFrame *dst, AVFrame *src)
-{
-    *dst = *src;
-    if (src->extended_data == src->data)
-        dst->extended_data = dst->data;
-    memset(src, 0, sizeof(*src));
-    get_frame_defaults(src);
-}
-
-int av_frame_is_writable(AVFrame *frame)
-{
-    int i, ret = 1;
-
-    /* assume non-refcounted frames are not writable */
-    if (!frame->buf[0])
-        return 0;
-
-    for (i = 0; i < FF_ARRAY_ELEMS(frame->buf); i++)
-        if (frame->buf[i])
-            ret &= !!av_buffer_is_writable(frame->buf[i]);
-    for (i = 0; i < frame->nb_extended_buf; i++)
-        ret &= !!av_buffer_is_writable(frame->extended_buf[i]);
-
-    return ret;
-}
-
-int av_frame_copy_props(AVFrame *dst, const AVFrame *src)
-{
-    return frame_copy_props(dst, src, 1);
-}
-
-AVBufferRef *av_frame_get_plane_buffer(AVFrame *frame, int plane)
-{
-    uint8_t *data;
-    int planes, i;
-
-    if (frame->nb_samples) {
-        int channels = frame->channels;
-        if (!channels)
-            return NULL;
-        CHECK_CHANNELS_CONSISTENCY(frame);
-        planes = av_sample_fmt_is_planar(frame->format) ? channels : 1;
-    } else
-        planes = 4;
-
-    if (plane < 0 || plane >= planes || !frame->extended_data[plane])
-        return NULL;
-    data = frame->extended_data[plane];
-
-    for (i = 0; i < FF_ARRAY_ELEMS(frame->buf) && frame->buf[i]; i++) {
-        AVBufferRef *buf = frame->buf[i];
-        if (data >= buf->data && data < buf->data + buf->size)
-            return buf;
-    }
-    for (i = 0; i < frame->nb_extended_buf; i++) {
-        AVBufferRef *buf = frame->extended_buf[i];
-        if (data >= buf->data && data < buf->data + buf->size)
-            return buf;
-    }
-    return NULL;
-}
-
-static int frame_copy_audio(AVFrame *dst, const AVFrame *src)
-{
-    int planar   = av_sample_fmt_is_planar(dst->format);
-    int channels = dst->channels;
-    int planes   = planar ? channels : 1;
-    int i;
-
-    if (dst->nb_samples     != src->nb_samples ||
-        dst->channels       != src->channels ||
-        dst->channel_layout != src->channel_layout)
-        return AVERROR(EINVAL);
-
-    CHECK_CHANNELS_CONSISTENCY(src);
-
-    for (i = 0; i < planes; i++)
-        if (!dst->extended_data[i] || !src->extended_data[i])
-            return AVERROR(EINVAL);
-
-    av_samples_copy(dst->extended_data, src->extended_data, 0, 0,
-                    dst->nb_samples, channels, dst->format);
-
-    return 0;
-}
-
-int av_frame_copy(AVFrame *dst, const AVFrame *src)
-{
-    if (dst->format != src->format || dst->format < 0)
-        return AVERROR(EINVAL);
-    else if (dst->nb_samples > 0 && dst->channel_layout)
-        return frame_copy_audio(dst, src);
-    return AVERROR(EINVAL);
 }
