+++ ext/at3_standalone/frame.h
@@ -34,94 +34,6 @@
 #include "samplefmt.h"
 #include "version.h"
 
-
-/**
- * @defgroup lavu_frame AVFrame
- * @ingroup lavu_data
- *
- * @{
- * AVFrame is an abstraction for reference-counted raw multimedia data.
- */
-
-enum AVFrameSideDataType {
-    /**
-     * The data is the AVPanScan struct defined in libavcodec.
-     */
-    AV_FRAME_DATA_PANSCAN,
-    /**
-     * ATSC A53 Part 4 Closed Captions.
-     * A53 CC bitstream is stored as uint8_t in AVFrameSideData.data.
-     * The number of bytes of CC data is AVFrameSideData.size.
-     */
-    AV_FRAME_DATA_A53_CC,
-    /**
-     * Stereoscopic 3d metadata.
-     * The data is the AVStereo3D struct defined in libavutil/stereo3d.h.
-     */
-    AV_FRAME_DATA_STEREO3D,
-    /**
-     * The data is the AVMatrixEncoding enum defined in libavutil/channel_layout.h.
-     */
-    AV_FRAME_DATA_MATRIXENCODING,
-    /**
-     * Metadata relevant to a downmix procedure.
-     * The data is the AVDownmixInfo struct defined in libavutil/downmix_info.h.
-     */
-    AV_FRAME_DATA_DOWNMIX_INFO,
-    /**
-     * ReplayGain information in the form of the AVReplayGain struct.
-     */
-    AV_FRAME_DATA_REPLAYGAIN,
-    /**
-     * This side data contains a 3x3 transformation matrix describing an affine
-     * transformation that needs to be applied to the frame for correct
-     * presentation.
-     *
-     * See libavutil/display.h for a detailed description of the data.
-     */
-    AV_FRAME_DATA_DISPLAYMATRIX,
-    /**
-     * Active Format Description data consisting of a single byte as specified
-     * in ETSI TS 101 154 using AVActiveFormatDescription enum.
-     */
-    AV_FRAME_DATA_AFD,
-    /**
-     * Motion vectors exported by some codecs (on demand through the export_mvs
-     * flag set in the libavcodec AVCodecContext flags2 option).
-     * The data is the AVMotionVector struct defined in
-     * libavutil/motion_vector.h.
-     */
-    AV_FRAME_DATA_MOTION_VECTORS,
-    /**
-     * Recommmends skipping the specified number of samples. This is exported
-     * only if the "skip_manual" AVOption is set in libavcodec.
-     * This has the same format as AV_PKT_DATA_SKIP_SAMPLES.
-     * @code
-     * u32le number of samples to skip from start of this packet
-     * u32le number of samples to skip from end of this packet
-     * u8    reason for start skip
-     * u8    reason for end   skip (0=padding silence, 1=convergence)
-     * @endcode
-     */
-    AV_FRAME_DATA_SKIP_SAMPLES,
-    /**
-     * This side data must be associated with an audio frame and corresponds to
-     * enum AVAudioServiceType defined in avcodec.h.
-     */
-    AV_FRAME_DATA_AUDIO_SERVICE_TYPE,
-    /**
-     * Mastering display metadata associated with a video frame. The payload is
-     * an AVMasteringDisplayMetadata type and contains information about the
-     * mastering display color volume.
-     */
-    AV_FRAME_DATA_MASTERING_DISPLAY_METADATA,
-    /**
-     * The GOP timecode in 25 bit timecode format. Data format is 64-bit integer.
-     * This is set on the first frame of a GOP that has a temporal reference of 0.
-     */
-    AV_FRAME_DATA_GOP_TIMECODE
-};
-
 /**
  * This structure describes decoded (raw) audio or video data.
  *
@@ -350,28 +262,22 @@ typedef struct AVFrame {
  * The position of these field in the structure is not part of the ABI,
  * they should not be accessed directly outside libavutil.
  */
-int64_t av_frame_get_best_effort_timestamp(const AVFrame *frame);
-void    av_frame_set_best_effort_timestamp(AVFrame *frame, int64_t val);
-int64_t av_frame_get_pkt_duration         (const AVFrame *frame);
-void    av_frame_set_pkt_duration         (AVFrame *frame, int64_t val);
-int64_t av_frame_get_pkt_pos              (const AVFrame *frame);
-void    av_frame_set_pkt_pos              (AVFrame *frame, int64_t val);
-int64_t av_frame_get_channel_layout       (const AVFrame *frame);
-void    av_frame_set_channel_layout       (AVFrame *frame, int64_t val);
-int     av_frame_get_channels             (const AVFrame *frame);
-void    av_frame_set_channels             (AVFrame *frame, int     val);
-int     av_frame_get_sample_rate          (const AVFrame *frame);
-void    av_frame_set_sample_rate          (AVFrame *frame, int     val);
-int     av_frame_get_decode_error_flags   (const AVFrame *frame);
-void    av_frame_set_decode_error_flags   (AVFrame *frame, int     val);
-int     av_frame_get_pkt_size(const AVFrame *frame);
-void    av_frame_set_pkt_size(AVFrame *frame, int val);
-
-/**
- * Get the name of a colorspace.
- * @return a static string identifying the colorspace; can be NULL.
- */
-const char *av_get_colorspace_name(enum AVColorSpace val);
+int64_t av_frame_get_best_effort_timestamp(const AVFrame * frame);
+void    av_frame_set_best_effort_timestamp(AVFrame * frame, int64_t val);
+int64_t av_frame_get_pkt_duration(const AVFrame * frame);
+void    av_frame_set_pkt_duration(AVFrame * frame, int64_t val);
+int64_t av_frame_get_pkt_pos(const AVFrame * frame);
+void    av_frame_set_pkt_pos(AVFrame * frame, int64_t val);
+int64_t av_frame_get_channel_layout(const AVFrame * frame);
+void    av_frame_set_channel_layout(AVFrame * frame, int64_t val);
+int     av_frame_get_channels(const AVFrame * frame);
+void    av_frame_set_channels(AVFrame * frame, int     val);
+int     av_frame_get_sample_rate(const AVFrame * frame);
+void    av_frame_set_sample_rate(AVFrame * frame, int     val);
+int     av_frame_get_decode_error_flags(const AVFrame * frame);
+void    av_frame_set_decode_error_flags(AVFrame * frame, int     val);
+int     av_frame_get_pkt_size(const AVFrame * frame);
+void    av_frame_set_pkt_size(AVFrame * frame, int val);
 
 /**
  * Allocate an AVFrame and set its fields to default values.  The resulting
@@ -395,36 +301,9 @@ AVFrame *av_frame_alloc(void);
 void av_frame_free(AVFrame **frame);
 
 /**
- * Set up a new reference to the data described by the source frame.
- *
- * Copy frame properties from src to dst and create a new reference for each
- * AVBufferRef from src.
- *
- * If src is not reference counted, new buffers are allocated and the data is
- * copied.
- *
- * @return 0 on success, a negative AVERROR on error
- */
-int av_frame_ref(AVFrame *dst, const AVFrame *src);
-
-/**
- * Create a new frame that references the same data as src.
- *
- * This is a shortcut for av_frame_alloc()+av_frame_ref().
- *
- * @return newly created AVFrame on success, NULL on error.
- */
-AVFrame *av_frame_clone(const AVFrame *src);
-
-/**
  * Unreference all the buffers referenced by frame and reset the frame fields.
  */
 void av_frame_unref(AVFrame *frame);
-
-/**
- * Move everything contained in src to dst and reset src.
- */
-void av_frame_move_ref(AVFrame *dst, AVFrame *src);
 
 /**
  * Allocate new buffer(s) for audio or video data.
@@ -444,53 +323,6 @@ void av_frame_move_ref(AVFrame *dst, AVFrame *src);
  * @return 0 on success, a negative AVERROR on error.
  */
 int av_frame_get_buffer(AVFrame *frame, int align);
-
-/**
- * Check if the frame data is writable.
- *
- * @return A positive value if the frame data is writable (which is true if and
- * only if each of the underlying buffers has only one reference, namely the one
- * stored in this frame). Return 0 otherwise.
- *
- * If 1 is returned the answer is valid until av_buffer_ref() is called on any
- * of the underlying AVBufferRefs (e.g. through av_frame_ref() or directly).
- *
- * @see av_frame_make_writable(), av_buffer_is_writable()
- */
-int av_frame_is_writable(AVFrame *frame);
-
-/**
- * Copy the frame data from src to dst.
- *
- * This function does not allocate anything, dst must be already initialized and
- * allocated with the same parameters as src.
- *
- * This function only copies the frame data (i.e. the contents of the data /
- * extended data arrays), not any other properties.
- *
- * @return >= 0 on success, a negative AVERROR on error.
- */
-int av_frame_copy(AVFrame *dst, const AVFrame *src);
-
-/**
- * Copy only "metadata" fields from src to dst.
- *
- * Metadata for the purpose of this function are those fields that do not affect
- * the data layout in the buffers.  E.g. pts, sample rate (for audio) or sample
- * aspect ratio (for video), but not width/height or channel layout.
- * Side data is also copied.
- */
-int av_frame_copy_props(AVFrame *dst, const AVFrame *src);
-
-/**
- * Get the buffer reference a given data plane is stored in.
- *
- * @param plane index of the data plane of interest in frame->extended_data.
- *
- * @return the buffer reference that contains the plane or NULL if the input
- * frame is not valid.
- */
-AVBufferRef *av_frame_get_plane_buffer(AVFrame *frame, int plane);
 
 /**
  * @}
