+++ ext/at3_standalone/mathematics.c
@@ -142,69 +142,3 @@ int64_t av_rescale_q(int64_t a, AVRational bq, AVRational cq)
 {
     return av_rescale_q_rnd(a, bq, cq, AV_ROUND_NEAR_INF);
 }
-
-int av_compare_ts(int64_t ts_a, AVRational tb_a, int64_t ts_b, AVRational tb_b)
-{
-    int64_t a = tb_a.num * (int64_t)tb_b.den;
-    int64_t b = tb_b.num * (int64_t)tb_a.den;
-    if ((FFABS(ts_a)|a|FFABS(ts_b)|b) <= INT_MAX)
-        return (ts_a*a > ts_b*b) - (ts_a*a < ts_b*b);
-    if (av_rescale_rnd(ts_a, a, b, AV_ROUND_DOWN) < ts_b)
-        return -1;
-    if (av_rescale_rnd(ts_b, b, a, AV_ROUND_DOWN) < ts_a)
-        return 1;
-    return 0;
-}
-
-int64_t av_compare_mod(uint64_t a, uint64_t b, uint64_t mod)
-{
-    int64_t c = (a - b) & (mod - 1);
-    if (c > (mod >> 1))
-        c -= mod;
-    return c;
-}
-
-int64_t av_rescale_delta(AVRational in_tb, int64_t in_ts,  AVRational fs_tb, int duration, int64_t *last, AVRational out_tb){
-    int64_t a, b, this;
-
-    av_assert0(in_ts != AV_NOPTS_VALUE);
-    av_assert0(duration >= 0);
-
-    if (*last == AV_NOPTS_VALUE || !duration || in_tb.num*(int64_t)out_tb.den <= out_tb.num*(int64_t)in_tb.den) {
-simple_round:
-        *last = av_rescale_q(in_ts, in_tb, fs_tb) + duration;
-        return av_rescale_q(in_ts, in_tb, out_tb);
-    }
-
-    a =  av_rescale_q_rnd(2*in_ts-1, in_tb, fs_tb, AV_ROUND_DOWN)   >>1;
-    b = (av_rescale_q_rnd(2*in_ts+1, in_tb, fs_tb, AV_ROUND_UP  )+1)>>1;
-    if (*last < 2*a - b || *last > 2*b - a)
-        goto simple_round;
-
-    this = av_clip64(*last, a, b);
-    *last = this + duration;
-
-    return av_rescale_q(this, fs_tb, out_tb);
-}
-
-int64_t av_add_stable(AVRational ts_tb, int64_t ts, AVRational inc_tb, int64_t inc)
-{
-    int64_t m, d;
-
-    if (inc != 1)
-        inc_tb = av_mul_q(inc_tb, (AVRational) {inc, 1});
-
-    m = inc_tb.num * (int64_t)ts_tb.den;
-    d = inc_tb.den * (int64_t)ts_tb.num;
-
-    if (m % d == 0)
-        return ts + m / d;
-    if (m < d)
-        return ts;
-
-    {
-        int64_t old = av_rescale_q(ts, ts_tb, inc_tb);
-        int64_t old_ts = av_rescale_q(old, inc_tb, ts_tb);
-        return av_rescale_q(old + 1, inc_tb, ts_tb) + (ts - old_ts);
-    }
-}
