+++ ext/at3_standalone/rational.c
@@ -97,11 +97,6 @@ AVRational av_add_q(AVRational b, AVRational c) {
     return b;
 }
 
-AVRational av_sub_q(AVRational b, AVRational c)
-{
-    return av_add_q(b, (AVRational) { -c.num, c.den });
-}
-
 AVRational av_d2q(double d, int max)
 {
     AVRational a;
@@ -121,178 +116,3 @@ AVRational av_d2q(double d, int max)
 
     return a;
 }
-
-int av_nearer_q(AVRational q, AVRational q1, AVRational q2)
-{
-    /* n/d is q, a/b is the median between q1 and q2 */
-    int64_t a = q1.num * (int64_t)q2.den + q2.num * (int64_t)q1.den;
-    int64_t b = 2 * (int64_t)q1.den * q2.den;
-
-    /* rnd_up(a*d/b) > n => a*d/b > n */
-    int64_t x_up = av_rescale_rnd(a, q.den, b, AV_ROUND_UP);
-
-    /* rnd_down(a*d/b) < n => a*d/b < n */
-    int64_t x_down = av_rescale_rnd(a, q.den, b, AV_ROUND_DOWN);
-
-    return ((x_up > q.num) - (x_down < q.num)) * av_cmp_q(q2, q1);
-}
-
-int av_find_nearest_q_idx(AVRational q, const AVRational* q_list)
-{
-    int i, nearest_q_idx = 0;
-    for (i = 0; q_list[i].den; i++)
-        if (av_nearer_q(q, q_list[i], q_list[nearest_q_idx]) > 0)
-            nearest_q_idx = i;
-
-    return nearest_q_idx;
-}
-
-uint32_t av_q2intfloat(AVRational q) {
-    int64_t n;
-    int shift;
-    int sign = 0;
-
-    if (q.den < 0) {
-        q.den *= -1;
-        q.num *= -1;
-    }
-    if (q.num < 0) {
-        q.num *= -1;
-        sign = 1;
-    }
-
-    if (!q.num && !q.den) return 0xFFC00000;
-    if (!q.num) return 0;
-    if (!q.den) return 0x7F800000 | (q.num & 0x80000000);
-
-    shift = 23 + av_log2(q.den) - av_log2(q.num);
-    if (shift >= 0) n = av_rescale(q.num, 1LL<<shift, q.den);
-    else            n = av_rescale(q.num, 1, ((int64_t)q.den) << -shift);
-
-    shift -= n >= (1<<24);
-    shift += n <  (1<<23);
-
-    if (shift >= 0) n = av_rescale(q.num, 1LL<<shift, q.den);
-    else            n = av_rescale(q.num, 1, ((int64_t)q.den) << -shift);
-
-    av_assert1(n <  (1<<24));
-    av_assert1(n >= (1<<23));
-
-    return sign<<31 | (150-shift)<<23 | (n - (1<<23));
-}
-
-#ifdef TEST
-
-#include "integer.h"
-
-int main(void)
-{
-    AVRational a,b,r;
-    int i,j,k;
-    static const int64_t numlist[] = {
-        INT64_MIN, INT64_MIN+1, INT64_MAX, INT32_MIN, INT32_MAX, 1,0,-1,
-        123456789, INT32_MAX-1, INT32_MAX+1LL, UINT32_MAX-1, UINT32_MAX, UINT32_MAX+1LL
-    };
-
-    for (a.num = -2; a.num <= 2; a.num++) {
-        for (a.den = -2; a.den <= 2; a.den++) {
-            for (b.num = -2; b.num <= 2; b.num++) {
-                for (b.den = -2; b.den <= 2; b.den++) {
-                    int c = av_cmp_q(a,b);
-                    double d = av_q2d(a) == av_q2d(b) ?
-                               0 : (av_q2d(a) - av_q2d(b));
-                    if (d > 0)       d = 1;
-                    else if (d < 0)  d = -1;
-                    else if (d != d) d = INT_MIN;
-                    if (c != d)
-                        av_log(NULL, AV_LOG_ERROR, "%d/%d %d/%d, %d %f\n", a.num,
-                               a.den, b.num, b.den, c,d);
-                    r = av_sub_q(av_add_q(b,a), b);
-                    if(b.den && (r.num*a.den != a.num*r.den || !r.num != !a.num || !r.den != !a.den))
-                        av_log(NULL, AV_LOG_ERROR, "%d/%d ", r.num, r.den);
-                }
-            }
-        }
-    }
-
-    for (i = 0; i < FF_ARRAY_ELEMS(numlist); i++) {
-        int64_t a = numlist[i];
-
-        for (j = 0; j < FF_ARRAY_ELEMS(numlist); j++) {
-            int64_t b = numlist[j];
-            if (b<=0)
-                continue;
-            for (k = 0; k < FF_ARRAY_ELEMS(numlist); k++) {
-                int64_t c = numlist[k];
-                int64_t res;
-                AVInteger ai;
-
-                if (c<=0)
-                    continue;
-                res = av_rescale_rnd(a,b,c, AV_ROUND_ZERO);
-
-                ai = av_mul_i(av_int2i(a), av_int2i(b));
-                ai = av_div_i(ai, av_int2i(c));
-
-                if (av_cmp_i(ai, av_int2i(INT64_MAX)) > 0 && res == INT64_MIN)
-                    continue;
-                if (av_cmp_i(ai, av_int2i(INT64_MIN)) < 0 && res == INT64_MIN)
-                    continue;
-                if (av_cmp_i(ai, av_int2i(res)) == 0)
-                    continue;
-
-                // Special exception for INT64_MIN, remove this in case INT64_MIN is handled without off by 1 error
-                if (av_cmp_i(ai, av_int2i(res-1)) == 0 && a == INT64_MIN)
-                    continue;
-
-                av_log(NULL, AV_LOG_ERROR, "%"PRId64" * %"PRId64" / %"PRId64" = %"PRId64" or %"PRId64"\n", a,b,c, res, av_i2int(ai));
-            }
-        }
-    }
-
-    for (a.num = 1; a.num <= 10; a.num++) {
-        for (a.den = 1; a.den <= 10; a.den++) {
-            if (av_gcd(a.num, a.den) > 1)
-                continue;
-            for (b.num = 1; b.num <= 10; b.num++) {
-                for (b.den = 1; b.den <= 10; b.den++) {
-                    int start;
-                    if (av_gcd(b.num, b.den) > 1)
-                        continue;
-                    if (av_cmp_q(b, a) < 0)
-                        continue;
-                    for (start = 0; start < 10 ; start++) {
-                        int acc= start;
-                        int i;
-
-                        for (i = 0; i<100; i++) {
-                            int exact = start + av_rescale_q(i+1, b, a);
-                            acc = av_add_stable(a, acc, b, 1);
-                            if (FFABS(acc - exact) > 2) {
-                                av_log(NULL, AV_LOG_ERROR, "%d/%d %d/%d, %d %d\n", a.num,
-                                       a.den, b.num, b.den, acc, exact);
-                                return 1;
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    for (a.den = 1; a.den < 0x100000000U/3; a.den*=3) {
-        for (a.num = -1; a.num < (1<<27); a.num += 1 + a.num/100) {
-            float f  = av_int2float(av_q2intfloat(a));
-            float f2 = av_q2d(a);
-            if (fabs(f - f2) > fabs(f)/5000000) {
-                av_log(NULL, AV_LOG_ERROR, "%d/%d %f %f\n", a.num,
-                       a.den, f, f2);
-                return 1;
-            }
-
-        }
-    }
-
-    return 0;
-}
-#endif
