+++ ext/at3_standalone/utils.c
@@ -33,10 +33,8 @@
 #include "frame.h"
 #include "internal.h"
 #include "mathematics.h"
-#include "avstring.h"
 #include "samplefmt.h"
 #include "avcodec.h"
-#include "opt.h"
 #include "internal.h"
 #include "bytestream.h"
 #include "version.h"
@@ -65,18 +63,6 @@ void av_fast_padded_malloc(void *ptr, unsigned int *size, size_t min_size)
         memset(*p + min_size, 0, AV_INPUT_BUFFER_PADDING_SIZE);
 }
 
-void av_fast_padded_mallocz(void *ptr, unsigned int *size, size_t min_size)
-{
-    uint8_t **p = ptr;
-    if (min_size > SIZE_MAX - AV_INPUT_BUFFER_PADDING_SIZE) {
-        av_freep(p);
-        *size = 0;
-        return;
-    }
-    if (!ff_fast_malloc(p, size, min_size + AV_INPUT_BUFFER_PADDING_SIZE, 1))
-        memset(*p, 0, min_size + AV_INPUT_BUFFER_PADDING_SIZE);
-}
-
 /* encoder management */
 static AVCodec *first_avcodec = NULL;
 static AVCodec **last_avcodec = &first_avcodec;
@@ -98,11 +84,6 @@ static av_cold void avcodec_init(void)
     initialized = 1;
 }
 
-int av_codec_is_encoder(const AVCodec *codec)
-{
-    return codec && (codec->encode_sub || codec->encode2);
-}
-
 int av_codec_is_decoder(const AVCodec *codec)
 {
     return codec && codec->decode;
@@ -114,42 +95,6 @@ unsigned avcodec_get_edge_width(void)
     return EDGE_WIDTH;
 }
 #endif
-
-int avcodec_fill_audio_frame(AVFrame *frame, int nb_channels,
-                             enum AVSampleFormat sample_fmt, const uint8_t *buf,
-                             int buf_size, int align)
-{
-    int ch, planar, needed_size, ret = 0;
-
-    needed_size = av_samples_get_buffer_size(NULL, nb_channels,
-                                             frame->nb_samples, sample_fmt,
-                                             align);
-    if (buf_size < needed_size)
-        return AVERROR(EINVAL);
-
-    planar = av_sample_fmt_is_planar(sample_fmt);
-    if (planar && nb_channels > AV_NUM_DATA_POINTERS) {
-        if (!(frame->extended_data = av_mallocz_array(nb_channels,
-                                                sizeof(*frame->extended_data))))
-            return AVERROR(ENOMEM);
-    } else {
-        frame->extended_data = frame->data;
-    }
-
-    if ((ret = av_samples_fill_arrays(frame->extended_data, &frame->linesize[0],
-                                      (uint8_t *)(intptr_t)buf, nb_channels, frame->nb_samples,
-                                      sample_fmt, align)) < 0) {
-        if (frame->extended_data != frame->data)
-            av_freep(&frame->extended_data);
-        return ret;
-    }
-    if (frame->extended_data != frame->data) {
-        for (ch = 0; ch < AV_NUM_DATA_POINTERS; ch++)
-            frame->data[ch] = frame->extended_data[ch];
-    }
-
-    return ret;
-}
 
 static int update_frame_pool(AVCodecContext *avctx, AVFrame *frame)
 {
@@ -309,61 +254,10 @@ int ff_get_buffer(AVCodecContext *avctx, AVFrame *frame, int flags)
 	return ret;
 }
 
-int avcodec_default_execute(AVCodecContext *c, int (*func)(AVCodecContext *c2, void *arg2), void *arg, int *ret, int count, int size)
-{
-    int i;
-
-    for (i = 0; i < count; i++) {
-        int r = func(c, (char *)arg + i * size);
-        if (ret)
-            ret[i] = r;
-    }
-    return 0;
-}
-
-int avcodec_default_execute2(AVCodecContext *c, int (*func)(AVCodecContext *c2, void *arg2, int jobnr, int threadnr), void *arg, int *ret, int count)
-{
-    int i;
-
-    for (i = 0; i < count; i++) {
-        int r = func(c, arg, i, 0);
-        if (ret)
-            ret[i] = r;
-    }
-    return 0;
-}
-
 MAKE_ACCESSORS(AVCodecContext, codec, AVRational, pkt_timebase)
 MAKE_ACCESSORS(AVCodecContext, codec, int, seek_preroll)
 
-unsigned av_codec_get_codec_properties(const AVCodecContext *codec)
-{
-    return codec->properties;
-}
-
-static int64_t get_bit_rate(AVCodecContext *ctx)
-{
-    int64_t bit_rate;
-    int bits_per_sample;
-
-    switch (ctx->codec_type) {
-    case AVMEDIA_TYPE_DATA:
-    case AVMEDIA_TYPE_SUBTITLE:
-    case AVMEDIA_TYPE_ATTACHMENT:
-        bit_rate = ctx->bit_rate;
-        break;
-    case AVMEDIA_TYPE_AUDIO:
-        bits_per_sample = av_get_bits_per_sample(ctx->codec_id);
-        bit_rate = bits_per_sample ? ctx->sample_rate * (int64_t)ctx->channels * bits_per_sample : ctx->bit_rate;
-        break;
-    default:
-        bit_rate = 0;
-        break;
-    }
-    return bit_rate;
-}
-
-int attribute_align_arg ff_codec_open2_recursive(AVCodecContext *avctx, const AVCodec *codec, AVDictionary **options)
+int attribute_align_arg ff_codec_open2_recursive(AVCodecContext *avctx, const AVCodec *codec, void **options)
 {
     int ret = 0;
 
@@ -375,7 +269,7 @@ int attribute_align_arg ff_codec_open2_recursive(AVCodecContext *avctx, const AV
     return ret;
 }
 
-int attribute_align_arg avcodec_open2(AVCodecContext *avctx, const AVCodec *codec, AVDictionary **options)
+int attribute_align_arg avcodec_open2(AVCodecContext *avctx, const AVCodec *codec, void **options)
 {
     int ret = 0;
 
@@ -424,17 +318,11 @@ int attribute_align_arg avcodec_open2(AVCodecContext *avctx, const AVCodec *code
             }
             if (codec->priv_class) {
                 *(const AVClass **)avctx->priv_data = codec->priv_class;
-                av_opt_set_defaults(avctx->priv_data);
             }
         }
     } else {
         avctx->priv_data = NULL;
     }
-
-    /* if the decoder init function was already called previously,
-     * free the already allocated subtitle_header before overwriting it */
-    if (av_codec_is_decoder(codec))
-        av_freep(&avctx->subtitle_header);
 
     if (avctx->channels > FF_SANE_NB_CHANNELS) {
         ret = AVERROR(EINVAL);
@@ -468,11 +356,6 @@ int attribute_align_arg avcodec_open2(AVCodecContext *avctx, const AVCodec *code
 #endif
 
 
-    avctx->pts_correction_num_faulty_pts =
-    avctx->pts_correction_num_faulty_dts = 0;
-    avctx->pts_correction_last_pts =
-    avctx->pts_correction_last_dts = INT64_MIN;
-
     if (avctx->codec->init) {
         ret = avctx->codec->init(avctx);
         if (ret < 0) {
@@ -488,8 +371,6 @@ int attribute_align_arg avcodec_open2(AVCodecContext *avctx, const AVCodec *code
 #endif
 
     if (av_codec_is_decoder(avctx->codec)) {
-        if (!avctx->bit_rate)
-            avctx->bit_rate = get_bit_rate(avctx);
         /* validate channel layout from the decoder */
         if (avctx->channel_layout) {
             int channels = av_get_channel_layout_nb_channels(avctx->channel_layout);
@@ -527,10 +408,6 @@ free_and_end:
     if (avctx->codec &&
         (avctx->codec->caps_internal & FF_CODEC_CAP_INIT_CLEANUP))
         avctx->codec->close(avctx);
-
-    if (codec->priv_class && codec->priv_data_size)
-        av_opt_free(avctx->priv_data);
-    av_opt_free(avctx);
 
 #if FF_API_CODED_FRAME
 FF_DISABLE_DEPRECATION_WARNINGS
@@ -613,86 +490,9 @@ av_cold int avcodec_close(AVCodecContext *avctx)
         av_freep(&avctx->internal->pool);
         av_freep(&avctx->internal);
     }
-
-    if (avctx->priv_data && avctx->codec && avctx->codec->priv_class)
-        av_opt_free(avctx->priv_data);
-    av_opt_free(avctx);
     av_freep(&avctx->priv_data);
-    if (av_codec_is_encoder(avctx->codec)) {
-        av_freep(&avctx->extradata);
-#if FF_API_CODED_FRAME
-FF_DISABLE_DEPRECATION_WARNINGS
-        av_frame_free(&avctx->coded_frame);
-FF_ENABLE_DEPRECATION_WARNINGS
-#endif
-    }
     avctx->codec = NULL;
     return 0;
-}
-
-static enum AVCodecID remap_deprecated_codec_id(enum AVCodecID id)
-{
-    switch(id){
-        //This is for future deprecatec codec ids, its empty since
-        //last major bump but will fill up again over time, please don't remove it
-        default                                         : return id;
-    }
-}
-
-static AVCodec *find_encdec(enum AVCodecID id, int encoder)
-{
-    AVCodec *p, *experimental = NULL;
-    p = first_avcodec;
-    id= remap_deprecated_codec_id(id);
-    while (p) {
-        if ((encoder ? av_codec_is_encoder(p) : av_codec_is_decoder(p)) &&
-            p->id == id) {
-            if (p->capabilities & AV_CODEC_CAP_EXPERIMENTAL && !experimental) {
-                experimental = p;
-            } else
-                return p;
-        }
-        p = p->next;
-    }
-    return experimental;
-}
-
-AVCodec *avcodec_find_encoder(enum AVCodecID id)
-{
-    return find_encdec(id, 1);
-}
-
-AVCodec *avcodec_find_encoder_by_name(const char *name)
-{
-    AVCodec *p;
-    if (!name)
-        return NULL;
-    p = first_avcodec;
-    while (p) {
-        if (av_codec_is_encoder(p) && strcmp(name, p->name) == 0)
-            return p;
-        p = p->next;
-    }
-    return NULL;
-}
-
-AVCodec *avcodec_find_decoder(enum AVCodecID id)
-{
-    return find_encdec(id, 0);
-}
-
-AVCodec *avcodec_find_decoder_by_name(const char *name)
-{
-    AVCodec *p;
-    if (!name)
-        return NULL;
-    p = first_avcodec;
-    while (p) {
-        if (av_codec_is_decoder(p) && strcmp(name, p->name) == 0)
-            return p;
-        p = p->next;
-    }
-    return NULL;
 }
 
 void avcodec_flush_buffers(AVCodecContext *avctx)
@@ -700,380 +500,8 @@ void avcodec_flush_buffers(AVCodecContext *avctx)
     if (avctx->codec->flush)
         avctx->codec->flush(avctx);
 
-    avctx->pts_correction_last_pts =
-    avctx->pts_correction_last_dts = INT64_MIN;
-
     if (!avctx->refcounted_frames)
         av_frame_unref(avctx->internal->to_free);
-}
-
-int av_get_exact_bits_per_sample(enum AVCodecID codec_id)
-{
-    switch (codec_id) {
-    case AV_CODEC_ID_8SVX_EXP:
-    case AV_CODEC_ID_8SVX_FIB:
-    case AV_CODEC_ID_ADPCM_CT:
-    case AV_CODEC_ID_ADPCM_IMA_APC:
-    case AV_CODEC_ID_ADPCM_IMA_EA_SEAD:
-    case AV_CODEC_ID_ADPCM_IMA_OKI:
-    case AV_CODEC_ID_ADPCM_IMA_WS:
-    case AV_CODEC_ID_ADPCM_G722:
-    case AV_CODEC_ID_ADPCM_YAMAHA:
-    case AV_CODEC_ID_ADPCM_AICA:
-        return 4;
-    case AV_CODEC_ID_DSD_LSBF:
-    case AV_CODEC_ID_DSD_MSBF:
-    case AV_CODEC_ID_DSD_LSBF_PLANAR:
-    case AV_CODEC_ID_DSD_MSBF_PLANAR:
-    case AV_CODEC_ID_PCM_ALAW:
-    case AV_CODEC_ID_PCM_MULAW:
-    case AV_CODEC_ID_PCM_S8:
-    case AV_CODEC_ID_PCM_S8_PLANAR:
-    case AV_CODEC_ID_PCM_U8:
-    case AV_CODEC_ID_PCM_ZORK:
-    case AV_CODEC_ID_SDX2_DPCM:
-        return 8;
-    case AV_CODEC_ID_PCM_S16BE:
-    case AV_CODEC_ID_PCM_S16BE_PLANAR:
-    case AV_CODEC_ID_PCM_S16LE:
-    case AV_CODEC_ID_PCM_S16LE_PLANAR:
-    case AV_CODEC_ID_PCM_U16BE:
-    case AV_CODEC_ID_PCM_U16LE:
-        return 16;
-    case AV_CODEC_ID_PCM_S24DAUD:
-    case AV_CODEC_ID_PCM_S24BE:
-    case AV_CODEC_ID_PCM_S24LE:
-    case AV_CODEC_ID_PCM_S24LE_PLANAR:
-    case AV_CODEC_ID_PCM_U24BE:
-    case AV_CODEC_ID_PCM_U24LE:
-        return 24;
-    case AV_CODEC_ID_PCM_S32BE:
-    case AV_CODEC_ID_PCM_S32LE:
-    case AV_CODEC_ID_PCM_S32LE_PLANAR:
-    case AV_CODEC_ID_PCM_U32BE:
-    case AV_CODEC_ID_PCM_U32LE:
-    case AV_CODEC_ID_PCM_F32BE:
-    case AV_CODEC_ID_PCM_F32LE:
-        return 32;
-    case AV_CODEC_ID_PCM_F64BE:
-    case AV_CODEC_ID_PCM_F64LE:
-        return 64;
-    default:
-        return 0;
-    }
-}
-
-enum AVCodecID av_get_pcm_codec(enum AVSampleFormat fmt, int be)
-{
-    static const enum AVCodecID map[AV_SAMPLE_FMT_NB][2] = {
-        [AV_SAMPLE_FMT_U8  ] = { AV_CODEC_ID_PCM_U8,    AV_CODEC_ID_PCM_U8    },
-        [AV_SAMPLE_FMT_S16 ] = { AV_CODEC_ID_PCM_S16LE, AV_CODEC_ID_PCM_S16BE },
-        [AV_SAMPLE_FMT_S32 ] = { AV_CODEC_ID_PCM_S32LE, AV_CODEC_ID_PCM_S32BE },
-        [AV_SAMPLE_FMT_FLT ] = { AV_CODEC_ID_PCM_F32LE, AV_CODEC_ID_PCM_F32BE },
-        [AV_SAMPLE_FMT_DBL ] = { AV_CODEC_ID_PCM_F64LE, AV_CODEC_ID_PCM_F64BE },
-        [AV_SAMPLE_FMT_U8P ] = { AV_CODEC_ID_PCM_U8,    AV_CODEC_ID_PCM_U8    },
-        [AV_SAMPLE_FMT_S16P] = { AV_CODEC_ID_PCM_S16LE, AV_CODEC_ID_PCM_S16BE },
-        [AV_SAMPLE_FMT_S32P] = { AV_CODEC_ID_PCM_S32LE, AV_CODEC_ID_PCM_S32BE },
-        [AV_SAMPLE_FMT_FLTP] = { AV_CODEC_ID_PCM_F32LE, AV_CODEC_ID_PCM_F32BE },
-        [AV_SAMPLE_FMT_DBLP] = { AV_CODEC_ID_PCM_F64LE, AV_CODEC_ID_PCM_F64BE },
-    };
-    if (fmt < 0 || fmt >= AV_SAMPLE_FMT_NB)
-        return AV_CODEC_ID_NONE;
-    if (be < 0 || be > 1)
-        be = AV_NE(1, 0);
-    return map[fmt][be];
-}
-
-int av_get_bits_per_sample(enum AVCodecID codec_id)
-{
-    switch (codec_id) {
-    case AV_CODEC_ID_ADPCM_SBPRO_2:
-        return 2;
-    case AV_CODEC_ID_ADPCM_SBPRO_3:
-        return 3;
-    case AV_CODEC_ID_ADPCM_SBPRO_4:
-    case AV_CODEC_ID_ADPCM_IMA_WAV:
-    case AV_CODEC_ID_ADPCM_IMA_QT:
-    case AV_CODEC_ID_ADPCM_SWF:
-    case AV_CODEC_ID_ADPCM_MS:
-        return 4;
-    default:
-        return av_get_exact_bits_per_sample(codec_id);
-    }
-}
-
-int av_get_audio_frame_duration(AVCodecContext *avctx, int frame_bytes)
-{
-    int id, sr, ch, ba, tag, bps;
-
-    id  = avctx->codec_id;
-    sr  = avctx->sample_rate;
-    ch  = avctx->channels;
-    ba  = avctx->block_align;
-    tag = avctx->codec_tag;
-    bps = av_get_exact_bits_per_sample(avctx->codec_id);
-
-    /* codecs with an exact constant bits per sample */
-    if (bps > 0 && ch > 0 && frame_bytes > 0 && ch < 32768 && bps < 32768)
-        return (frame_bytes * 8LL) / (bps * ch);
-    bps = avctx->bits_per_coded_sample;
-
-    /* codecs with a fixed packet duration */
-    switch (id) {
-    case AV_CODEC_ID_ADPCM_ADX:    return   32;
-    case AV_CODEC_ID_ADPCM_IMA_QT: return   64;
-    case AV_CODEC_ID_ADPCM_EA_XAS: return  128;
-    case AV_CODEC_ID_AMR_NB:
-    case AV_CODEC_ID_EVRC:
-    case AV_CODEC_ID_GSM:
-    case AV_CODEC_ID_QCELP:
-    case AV_CODEC_ID_RA_288:       return  160;
-    case AV_CODEC_ID_AMR_WB:
-    case AV_CODEC_ID_GSM_MS:       return  320;
-    case AV_CODEC_ID_MP1:          return  384;
-    case AV_CODEC_ID_ATRAC1:       return  512;
-    case AV_CODEC_ID_ATRAC3:       return 1024;
-    case AV_CODEC_ID_ATRAC3P:      return 2048;
-    case AV_CODEC_ID_MP2:
-    case AV_CODEC_ID_MUSEPACK7:    return 1152;
-    case AV_CODEC_ID_AC3:          return 1536;
-    }
-
-    if (sr > 0) {
-        /* calc from sample rate */
-        if (id == AV_CODEC_ID_TTA)
-            return 256 * sr / 245;
-
-        if (ch > 0) {
-            /* calc from sample rate and channels */
-            if (id == AV_CODEC_ID_BINKAUDIO_DCT)
-                return (480 << (sr / 22050)) / ch;
-        }
-    }
-
-    if (ba > 0) {
-        /* calc from block_align */
-        if (id == AV_CODEC_ID_SIPR) {
-            switch (ba) {
-            case 20: return 160;
-            case 19: return 144;
-            case 29: return 288;
-            case 37: return 480;
-            }
-        } else if (id == AV_CODEC_ID_ILBC) {
-            switch (ba) {
-            case 38: return 160;
-            case 50: return 240;
-            }
-        }
-    }
-
-    if (frame_bytes > 0) {
-        /* calc from frame_bytes only */
-        if (id == AV_CODEC_ID_TRUESPEECH)
-            return 240 * (frame_bytes / 32);
-        if (id == AV_CODEC_ID_NELLYMOSER)
-            return 256 * (frame_bytes / 64);
-        if (id == AV_CODEC_ID_RA_144)
-            return 160 * (frame_bytes / 20);
-        if (id == AV_CODEC_ID_G723_1)
-            return 240 * (frame_bytes / 24);
-
-        if (bps > 0) {
-            /* calc from frame_bytes and bits_per_coded_sample */
-            if (id == AV_CODEC_ID_ADPCM_G726)
-                return frame_bytes * 8 / bps;
-        }
-
-        if (ch > 0 && ch < INT_MAX/16) {
-            /* calc from frame_bytes and channels */
-            switch (id) {
-            case AV_CODEC_ID_ADPCM_AFC:
-                return frame_bytes / (9 * ch) * 16;
-            case AV_CODEC_ID_ADPCM_PSX:
-            case AV_CODEC_ID_ADPCM_DTK:
-                return frame_bytes / (16 * ch) * 28;
-            case AV_CODEC_ID_ADPCM_4XM:
-            case AV_CODEC_ID_ADPCM_IMA_ISS:
-                return (frame_bytes - 4 * ch) * 2 / ch;
-            case AV_CODEC_ID_ADPCM_IMA_SMJPEG:
-                return (frame_bytes - 4) * 2 / ch;
-            case AV_CODEC_ID_ADPCM_IMA_AMV:
-                return (frame_bytes - 8) * 2 / ch;
-            case AV_CODEC_ID_ADPCM_THP:
-            case AV_CODEC_ID_ADPCM_THP_LE:
-                if (avctx->extradata)
-                    return frame_bytes * 14 / (8 * ch);
-                break;
-            case AV_CODEC_ID_ADPCM_XA:
-                return (frame_bytes / 128) * 224 / ch;
-            case AV_CODEC_ID_INTERPLAY_DPCM:
-                return (frame_bytes - 6 - ch) / ch;
-            case AV_CODEC_ID_ROQ_DPCM:
-                return (frame_bytes - 8) / ch;
-            case AV_CODEC_ID_XAN_DPCM:
-                return (frame_bytes - 2 * ch) / ch;
-            case AV_CODEC_ID_MACE3:
-                return 3 * frame_bytes / ch;
-            case AV_CODEC_ID_MACE6:
-                return 6 * frame_bytes / ch;
-            case AV_CODEC_ID_PCM_LXF:
-                return 2 * (frame_bytes / (5 * ch));
-            case AV_CODEC_ID_IAC:
-            case AV_CODEC_ID_IMC:
-                return 4 * frame_bytes / ch;
-            }
-
-            if (tag) {
-                /* calc from frame_bytes, channels, and codec_tag */
-                if (id == AV_CODEC_ID_SOL_DPCM) {
-                    if (tag == 3)
-                        return frame_bytes / ch;
-                    else
-                        return frame_bytes * 2 / ch;
-                }
-            }
-
-            if (ba > 0) {
-                /* calc from frame_bytes, channels, and block_align */
-                int blocks = frame_bytes / ba;
-                switch (avctx->codec_id) {
-                case AV_CODEC_ID_ADPCM_IMA_WAV:
-                    if (bps < 2 || bps > 5)
-                        return 0;
-                    return blocks * (1 + (ba - 4 * ch) / (bps * ch) * 8);
-                case AV_CODEC_ID_ADPCM_IMA_DK3:
-                    return blocks * (((ba - 16) * 2 / 3 * 4) / ch);
-                case AV_CODEC_ID_ADPCM_IMA_DK4:
-                    return blocks * (1 + (ba - 4 * ch) * 2 / ch);
-                case AV_CODEC_ID_ADPCM_IMA_RAD:
-                    return blocks * ((ba - 4 * ch) * 2 / ch);
-                case AV_CODEC_ID_ADPCM_MS:
-                    return blocks * (2 + (ba - 7 * ch) * 2 / ch);
-                }
-            }
-
-            if (bps > 0) {
-                /* calc from frame_bytes, channels, and bits_per_coded_sample */
-                switch (avctx->codec_id) {
-                case AV_CODEC_ID_PCM_DVD:
-                    if(bps<4)
-                        return 0;
-                    return 2 * (frame_bytes / ((bps * 2 / 8) * ch));
-                case AV_CODEC_ID_PCM_BLURAY:
-                    if(bps<4)
-                        return 0;
-                    return frame_bytes / ((FFALIGN(ch, 2) * bps) / 8);
-                case AV_CODEC_ID_S302M:
-                    return 2 * (frame_bytes / ((bps + 4) / 4)) / ch;
-                }
-            }
-        }
-    }
-
-    /* Fall back on using frame_size */
-    if (avctx->frame_size > 1 && frame_bytes)
-        return avctx->frame_size;
-
-    //For WMA we currently have no other means to calculate duration thus we
-    //do it here by assuming CBR, which is true for all known cases.
-    if (avctx->bit_rate>0 && frame_bytes>0 && avctx->sample_rate>0 && avctx->block_align>1) {
-        if (avctx->codec_id == AV_CODEC_ID_WMAV1 || avctx->codec_id == AV_CODEC_ID_WMAV2)
-            return  (frame_bytes * 8LL * avctx->sample_rate) / avctx->bit_rate;
-    }
-
-    return 0;
-}
-
-#if !HAVE_THREADS
-int ff_thread_init(AVCodecContext *s)
-{
-    return -1;
-}
-
-#endif
-
-unsigned int av_xiphlacing(unsigned char *s, unsigned int v)
-{
-    unsigned int n = 0;
-
-    while (v >= 0xff) {
-        *s++ = 0xff;
-        v -= 0xff;
-        n++;
-    }
-    *s = v;
-    n++;
-    return n;
-}
-
-int ff_match_2uint16(const uint16_t(*tab)[2], int size, int a, int b)
-{
-    int i;
-    for (i = 0; i < size && !(tab[i][0] == a && tab[i][1] == b); i++) ;
-    return i;
-}
-
-#if FF_API_MISSING_SAMPLE
-FF_DISABLE_DEPRECATION_WARNINGS
-void av_log_missing_feature(void *avc, const char *feature, int want_sample)
-{
-    av_log(avc, AV_LOG_WARNING, "%s is not implemented. Update your FFmpeg "
-            "version to the newest one from Git. If the problem still "
-            "occurs, it means that your file has a feature which has not "
-            "been implemented.\n", feature);
-    if(want_sample)
-        av_log_ask_for_sample(avc, NULL);
-}
-
-void av_log_ask_for_sample(void *avc, const char *msg, ...)
-{
-    va_list argument_list;
-
-    va_start(argument_list, msg);
-
-    if (msg)
-        av_vlog(avc, AV_LOG_WARNING, msg, argument_list);
-    av_log(avc, AV_LOG_WARNING, "If you want to help, upload a sample "
-            "of this file to ftp://upload.ffmpeg.org/incoming/ "
-            "and contact the ffmpeg-devel mailing list. (ffmpeg-devel@ffmpeg.org)\n");
-
-    va_end(argument_list);
-}
-FF_ENABLE_DEPRECATION_WARNINGS
-#endif /* FF_API_MISSING_SAMPLE */
-
-
-int av_lockmgr_register(int (*cb)(void **mutex, enum AVLockOp op))
-{
-    if (lockmgr_cb) {
-        // There is no good way to rollback a failure to destroy the
-        // mutex, so we ignore failures.
-        lockmgr_cb(&codec_mutex,    AV_LOCK_DESTROY);
-        lockmgr_cb(&avformat_mutex, AV_LOCK_DESTROY);
-        lockmgr_cb     = NULL;
-        codec_mutex    = NULL;
-        avformat_mutex = NULL;
-    }
-
-    if (cb) {
-        void *new_codec_mutex    = NULL;
-        void *new_avformat_mutex = NULL;
-        int err;
-        if (err = cb(&new_codec_mutex, AV_LOCK_CREATE)) {
-            return err > 0 ? AVERROR_UNKNOWN : err;
-        }
-        if (err = cb(&new_avformat_mutex, AV_LOCK_CREATE)) {
-            // Ignore failures to destroy the newly created mutex.
-            cb(&new_codec_mutex, AV_LOCK_DESTROY);
-            return err > 0 ? AVERROR_UNKNOWN : err;
-        }
-        lockmgr_cb     = cb;
-        codec_mutex    = new_codec_mutex;
-        avformat_mutex = new_avformat_mutex;
-    }
-
-    return 0;
 }
 
 int ff_lock_avcodec(AVCodecContext *log_ctx, const AVCodec *codec)
@@ -1106,80 +534,7 @@ int ff_unlock_avcodec(const AVCodec *codec)
     return 0;
 }
 
-int avpriv_lock_avformat(void)
-{
-    if (lockmgr_cb) {
-        if ((*lockmgr_cb)(&avformat_mutex, AV_LOCK_OBTAIN))
-            return -1;
-    }
-    return 0;
-}
-
-int avpriv_unlock_avformat(void)
-{
-    if (lockmgr_cb) {
-        if ((*lockmgr_cb)(&avformat_mutex, AV_LOCK_RELEASE))
-            return -1;
-    }
-    return 0;
-}
-
-unsigned int avpriv_toupper4(unsigned int x)
-{
-    return av_toupper(x & 0xFF) +
-          (av_toupper((x >>  8) & 0xFF) << 8)  +
-          (av_toupper((x >> 16) & 0xFF) << 16) +
-((unsigned)av_toupper((x >> 24) & 0xFF) << 24);
-}
-
 int avcodec_is_open(AVCodecContext *s)
 {
     return !!s->internal;
-}
-
-const uint8_t *avpriv_find_start_code(const uint8_t *av_restrict p,
-                                      const uint8_t *end,
-                                      uint32_t *av_restrict state)
-{
-    int i;
-
-    av_assert0(p <= end);
-    if (p >= end)
-        return end;
-
-    for (i = 0; i < 3; i++) {
-        uint32_t tmp = *state << 8;
-        *state = tmp + *(p++);
-        if (tmp == 0x100 || p == end)
-            return p;
-    }
-
-    while (p < end) {
-        if      (p[-1] > 1      ) p += 3;
-        else if (p[-2]          ) p += 2;
-        else if (p[-3]|(p[-1]-1)) p++;
-        else {
-            p++;
-            break;
-        }
-    }
-
-    p = FFMIN(p, end) - 4;
-    *state = AV_RB32(p);
-
-    return p + 4;
-}
-
-AVCPBProperties *av_cpb_properties_alloc(size_t *size)
-{
-    AVCPBProperties *props = av_mallocz(sizeof(AVCPBProperties));
-    if (!props)
-        return NULL;
-
-    if (size)
-        *size = sizeof(*props);
-
-    props->vbv_delay = UINT64_MAX;
-
-    return props;
 }
